sql-tool/requirements.txt
functions-framework==3.*
google-cloud-bigquery>=3.10.0
gunicorn
